requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
gradio>=4.0.0
pillow>=10.0.0
pyyaml>=6.0
huggingface-hub>=0.20.0
numpy>=1.24.0
safetensors>=0.4.0

